build on windows. debug iocp event actions.

+++ common/alloter/pool_alloter.cpp
@@ -105,7 +105,7 @@ void* PoolAlloter::ReFill(uint32_t size, uint32_t num) {
 void* PoolAlloter::ChunkAlloc(uint32_t size, uint32_t& nums) {
     char* res;
     uint32_t need_bytes = size * nums;
-    uint32_t left_bytes = _pool_end - _pool_start;
+    uint32_t left_bytes = uint32_t(_pool_end - _pool_start);
 
     //pool is enough
     if (left_bytes >= need_bytes) {
